requirements.txt
PyYAML==5.4.1
cbor2==5.2.0